feat: spots permission

+++ routes/spots.js
@@ -55,23 +55,38 @@ router.get('/:id', catchAsync(async (req, res) => {
 }));
 
 router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+    const { id } = req.params;
     const spot = await Spot.findById(req.params.id);
     if (!spot) {
         req.flash('error', 'Cannot edit find that spot!');
         return res.redirect('/spots');
+    }
+    if (!spot.author.equals(req.user._id)) {
+        req.flash('error','You do not have permission');
+        return res.redirect(`/spots/${id}`);
     }
     res.render('spots/edit', { spot })
 }));
 
 router.put('/:id', isLoggedIn, validateSpot, catchAsync(async (req, res) => {
     const { id } = req.params;
-    const spot = await Spot.findByIdAndUpdate(id, { ...req.body.spot });
+    const spot = await Spot.findById(id);
+    if (!spot.author.equals(req.user._id)) {
+        req.flash('error','You do not have permission');
+        return res.redirect(`/spots/${id}`);
+    }
+    await Spot.findByIdAndUpdate(id, { ...req.body.spot });
     req.flash('success', 'Successfully updated spot!');
     res.redirect(`/spots/${spot._id}`)
 }));
 
 router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const spot = await Spot.findById(id);
+    if (!spot.author.equals(req.user._id)) {
+        req.flash('error','You do not have permission');
+        return res.redirect(`/spots/${id}`);
+    }
     await Spot.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted spot!')
     res.redirect('/spots');
